requirements.txt
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Jupyter notebook environment
jupyter>=1.0.0
notebook>=6.4.0
jupyterlab>=3.2.0
ipywidgets>=7.6.0

# Web framework and API
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0

# Data validation and serialization
pydantic>=1.8.0
marshmallow>=3.14.0

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=5.4.0
configparser>=5.0.0

# HTTP requests and async
requests>=2.25.0
aiohttp>=3.7.0
httpx>=0.24.0

# Database and caching
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=3.5.0
pymongo>=3.12.0

# Task queue and background processing
celery>=5.2.0
flower>=1.0.0

# Machine learning and model serving
mlflow>=1.20.0
joblib>=1.1.0
pickle5>=0.0.11

# Geospatial and location
geopy>=2.2.0
folium>=0.12.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=3.4.0

# Monitoring and logging
prometheus-client>=0.11.0
structlog>=21.1.0

# Testing and development
pytest>=6.2.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0
black>=21.6b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0
pre-commit>=2.15.0

# Data processing and utilities
openpyxl>=3.0.0
xlrd>=2.0.0
python-dateutil>=2.8.0
pytz>=2021.1

# Financial and domain-specific

# Development and documentation
sphinx>=4.2.0
sphinx-rtd-theme>=0.5.0
